wufeng-med-qa/requirements.txt
streamlit==1.38.0
pypdf==3.17.4
faiss-cpu==1.8.0
numpy==1.24.4
pandas==1.5.3
sentence-transformers==2.6.1
PyYAML==6.0.2
python-dotenv==1.0.1
openai==0.28.1
dashscope